fix: Fetching user details
Corrects an error when fetching user information. We moved the
authentication endpoint to login.procore.com, however that incorrectly
handles the /vapid/me endpoint.

To fix, added an additonal option to the gem: `api_site`. This defaults
to the API gateway, api.procore.com which will correctly handle routing
/vapid/me.

Should the end user want to change either value (for usage with
sandboxes, for example), this will still support configuring the gem for
that purpose.

The configuration could look something like this:

```
  use OmniAuth::Builder do
    provider(
      :procore,
      ENV['PROCORE_KEY'],
      ENV['PROCORE_SECRET'],
      client_options: {
        site: ENV.fetch('PROCORE_AUTH_URL', 'https://login.procore.com')
        api_site: ENV.fetch('PROCORE_API_URL', 'https://api.procore.com')
      },
    )
  end
```